requirements.txt
pytest-playwright
openpyxl
pandas
tqdm
beautifulsoup4
aiohttp